refactor: deduplicate constants and flatten re-export chains (Phase 2)
- Move PROXY_API_TYPES + getProxyApiType to constants.ts (src/index.ts, auth.ts now import)
- Remove proxy-core.ts re-exports of UPSTREAM_SUCCESS/RawHeaders
- Remove proxy-logging.ts re-exports of UPSTREAM_SUCCESS/RawHeaders
- Direct imports from types.ts: log-helpers.ts, openai.ts
- Deduplicate isTarget() in scheduled.ts — reuse from targets-rule.ts
