+++ src/proxy/openai.ts
@@ -4,8 +4,9 @@ import fp from "fastify-plugin";
 import { getActiveProviders } from "../db/index.js";
 import { getSetting } from "../db/settings.js";
 import { decrypt } from "../utils/crypto.js";
-import { proxyGetRequest, createErrorFormatter, type RawHeaders, type ProxyErrorResponse } from "./proxy-core.js";
+import { proxyGetRequest, createErrorFormatter, type ProxyErrorResponse } from "./proxy-core.js";
 import type { ErrorKind } from "./proxy-core.js";
+import type { RawHeaders } from "./types.js";
 import { handleProxyRequest, type RouteHandlerDeps } from "./proxy-handler.js";
 import { createOrchestrator } from "./orchestrator.js";
 import { RetryRuleMatcher } from "./retry-rules.js";
