+++ src/proxy/proxy-logging.ts
@@ -7,9 +7,6 @@ import { MetricsExtractor } from "../metrics/metrics-extractor.js";
 import type { FastifyRequest } from "fastify";
 import type { ResilienceAttempt } from "./resilience.js";
 import type { TransportResult } from "./types.js";
-// Re-export shared constants & types from types.ts (canonical home)
-export { UPSTREAM_SUCCESS } from "./types.js";
-export type { RawHeaders } from "./types.js";
 
 // Internal imports from types.ts
 import { UPSTREAM_SUCCESS } from "./types.js";
